requirements.txt
datasets
torchmetrics
tokenizers
pytorch-lightning
torch-lr-finder
lightning
torch_tb_profiler